Restructures folders
