+++ babel-plugin/index.js
