+++ lib/transforms/create-new-definitions.js
