+++ lib/transforms/expand-references.js
