+++ lib/transforms/transform-prop-types.js
