+++ lib/utils/capitalize.js
