+++ lib/utils/get-component-name.js
