+++ lib/utils/get-prop-types-identifier-name.js
