+++ lib/utils/get-prop-types.js
