+++ lib/utils/is-enum-required.js
