+++ lib/utils/validate-props.js
