+++ test/generate-class.js
@@ -2,7 +2,7 @@ const fs = require('fs');
 const path = require('path');
 const test = require('ava');
 
-const generateClass = require('../source/generate-class');
+const generateClass = require('../lib');
 
 const funcComponentPath = path.join(
   __dirname,
