+++ test/get-prop-types-identifier-name.js
@@ -1,7 +1,7 @@
 const { parse } = require('@babel/parser');
 const test = require('ava');
 
-const getPropTypesIdentifierName = require('../source/utils/get-prop-types-identifier-name');
+const getPropTypesIdentifierName = require('../lib/utils/get-prop-types-identifier-name');
 
 test('With PropTypes import', t => {
   const syntaxTree = parse('import pt from "prop-types";', {
